llm_router/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
google-cloud-firestore==2.9.0
requests==2.28.2
